Fixed long-standing permissions bug on Linux and macOS.
Since Armory is a Blender addon that must be distributed through a .zip file, executable files will not work on macOS or Linux. This update adds code that automatically assigns the correct permissions every time the Armory addon is initialized. This allows users to simply install the addon on macOS and Linux and have it work. This has never been the case, and has always required extra setup on these platforms.

Unlike the last two pull requests, this one actually fixes the problem once and for all! The first time, untested code had issues on macOS. The second time, Xcode decided to undo all my work for some reason and I didn't notice until after I submitted the PR. This code is tested and working on macOS and Linux.